Move template utilities to common folder
We will use them for non-driver related things
